feat: code sepated into modules
